requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
numpy==1.26.4
pyarrow==12.0.1
lancedb==0.3.3
python-jose==3.3.0
python-multipart==0.0.6
passlib==1.7.4
python-docx==0.8.11
bcrypt==4.0.1
pydantic==2.3.0
python-dotenv==1.0.0 